Move collecting of available variables
